feat: separado componentes da pagina about em arquivos individuais
